refactor: extract receipt lookup hook
